feat: Finalize all code and assets for The Colosseum v1.4.0
